LayoutTests/fast/js/script-line-number-expected.txt

This test checks that line numbers are correctly reported for both inline scripts and inline event handlers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 16 is 16
PASS 19 is 19
PASS successfullyParsed is true

TEST COMPLETE
